refactor: separate code into modules
